src/test/java/learn/library/controller/LibraryControllerTest.java
package learn.library.controller;

import learn.library.service.LibraryImpl;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.model;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;

@WebMvcTest(LibraryController.class)
public class LibraryControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private LibraryImpl library;

    @MockBean
    private TestEntityManager testEntityManager;

    @Test
    public void getBooks() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders
                .get("/"))
                .andExpect(model().attributeExists("books"));
        verify(library).getBooks();
    }

    @Test
    void viewBook() throws Exception {
        mockMvc.perform(MockMvcRequestBuilders.get("/view?id=1"))
                .andExpect(view().name("view"));
    }

}